Add Account and Transaction folder and test

+++ src/main/java/cs203/g1t7/content/ContentForbiddenException.java
@@ -8,7 +8,7 @@ public class ContentForbiddenException extends RuntimeException {
 
     private static final long serialVersionUID = 1L;
 
-    public ContentForbiddenException(Long id) {
+    public ContentForbiddenException(Integer id) {
         super("Forbidden to view content " + id);
     }
     public ContentForbiddenException() {
