+++ src/main/java/cs203/g1t7/content/ContentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContentRepository extends JpaRepository <Content, Long> {
+public interface ContentRepository extends JpaRepository <Content, Integer> {
     List<Content> findByTitle(String title);
 }
