+++ src/main/java/cs203/g1t7/content/ContentService.java
@@ -4,13 +4,13 @@ import java.util.List;
 
 public interface ContentService {
     List<Content> listContents();
-    Content getContent(Long id);
+    Content getContent(Integer id);
     Content addContent(Content content);
-    Content updateContent(Long id, Content content);
+    Content updateContent(Integer id, Content content);
 
     /**
      * Change method's signature: do not return a value for delete operation
      * @param id
      */
-    void deleteContent(Long id);
+    void deleteContent(Integer id);
 }
